backend/requirements.txt
# Backend Requirements
# Django and extensions
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Database and GIS
psycopg2-binary>=2.9.0
Django-environ>=0.10.0

# APRS and packet processing
aprslib>=0.7.0
kiss>=7.0.0
pyserial>=3.5

# Weather radar processing
arm-pyart>=1.15.0
numpy>=1.24.0
matplotlib>=3.7.0
cartopy>=0.22.0
netCDF4>=1.6.0

# Geospatial
geopy>=2.3.0
shapely>=2.0.0

# WebSocket and async
redis>=4.5.0
celery>=5.3.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
pillow>=10.0.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-django>=4.5.0
